Moves Table, TabsComponent and Tags
