+++ amundsen_application/static/js/components/EditableSection/index.spec.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import TagInput from 'components/common/Tags/TagInput';
+import TagInput from 'components/Tags/TagInput';
 import { ResourceType } from 'interfaces/Resources';
 import EditableSection, { EditableSectionProps } from '.';
