+++ amundsen_application/static/js/components/Table/styles.scss
