+++ amundsen_application/static/js/components/TabsComponent/index.spec.tsx
