+++ amundsen_application/static/js/components/TabsComponent/styles.scss
