+++ amundsen_application/static/js/pages/BrowsePage/index.spec.tsx
@@ -6,7 +6,7 @@ import * as DocumentTitle from 'react-document-title';
 
 import { shallow } from 'enzyme';
 
-import TagsListContainer from 'components/common/Tags';
+import TagsListContainer from 'components/Tags';
 import { BrowsePage } from '.';
 
 describe('BrowsePage', () => {
