+++ amundsen_application/static/js/pages/BrowsePage/index.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import * as DocumentTitle from 'react-document-title';
 
-import TagsListContainer from 'components/common/Tags';
+import TagsListContainer from 'components/Tags';
 
 import { BROWSE_PAGE_DOCUMENT_TITLE } from './constants';
