+++ amundsen_application/static/js/pages/TableDetailPage/index.spec.tsx
@@ -10,7 +10,7 @@ import { getMockRouterProps } from 'fixtures/mockRouter';
 import { tableMetadata } from 'fixtures/metadata/table';
 
 import LoadingSpinner from 'components/LoadingSpinner';
-import TabsComponent from 'components/common/TabsComponent';
+import TabsComponent from 'components/TabsComponent';
 
 import { indexDashboardsEnabled } from 'config/config-utils';
 import { TableDetail, TableDetailProps, MatchProps } from '.';
